Restructured func_kk
